+++ air/src/trace/main_trace.rs
@@ -137,7 +137,7 @@ impl MainTrace {
 
     /// Returns a specific element from the hasher state at row i.
     pub fn decoder_hasher_state_element(&self, element: usize, i: RowIndex) -> Felt {
-        self.columns.get_column(DECODER_TRACE_OFFSET + HASHER_STATE_OFFSET + element)[i + 1]
+        self.columns.get_column(DECODER_TRACE_OFFSET + HASHER_STATE_OFFSET + element)[i]
     }
 
     /// Returns the current function hash (i.e., root) at row i.