+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Implemented `MastForest` merging (#1534)
 - Rename `EqHash` to `MastNodeFingerprint` and make it `pub` (#1539)
 - [BREAKING] `DYN` operation now expects a memory address pointing to the procedure hash (#1535)
+- [BREAKING] `DYNCALL` operation fixed, and now expects a memory address pointing to the procedure hash (#1535)
 
 
 #### Fixes